assingment_10/requirements.txt
langchain==0.1.0
openai==1.7.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
chromadb==0.4.21
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
